Add LibraryMenuService
