+++ src/app/views/designer/designer.component.ts
@@ -2,15 +2,8 @@ import { Component, OnDestroy, OnInit } from '@angular/core';
 import { Subscription } from 'rxjs';
 
 import { UrlService } from '../../services/url.service';
-import { SegmentPaths } from '../../segment-paths';
-
+import { LibraryMenuService } from '../shared/library/library-menu.service';
 import { MenuItem } from '../../models/menu-item.model';
-
-const TopLevelMenuItemName = {
-  components: 'Components',
-  colorsAndFonts: 'Colors, Fonts',
-  icons: 'Icons'
-};
 
 @Component({
   selector: 'app-designer',
@@ -23,31 +16,24 @@ export class DesignerComponent implements OnInit, OnDestroy {
 
   private subscription: Subscription = new Subscription();
 
-  constructor(private urlService: UrlService) { }
+  constructor(
+    private urlService: UrlService,
+    private libraryMenuService: LibraryMenuService
+  ) { }
 
   ngOnInit() {
-    this.topLevelMenuItems = this.getTopLevelMenuItems();
+    this.topLevelMenuItems = this.libraryMenuService.getTopLevelMenuItems();
     this.subscription.add(this.urlService.URL$
       .subscribe(() => {
-        console.log(this.urlService.getAbsolutePath());
+        if (this.libraryMenuService.updateSecondLevelMenu()) {
+          console.log(this.secondLevelMenuItems = this.libraryMenuService.getSecondLevelMenuItems());
+          // this.secondLevelMenuItems = this.libraryMenuService.getSecondLevelMenuItems();
+        }
       }));
   }
 
   ngOnDestroy() {
     this.subscription.unsubscribe();
-  }
-
-  private getTopLevelMenuItems(): MenuItem[] {
-    return [{
-      name: TopLevelMenuItemName.components,
-      link: `/${SegmentPaths.designer}/${SegmentPaths.components}`
-    }, {
-      name: TopLevelMenuItemName.colorsAndFonts,
-      link: `/${SegmentPaths.designer}/${SegmentPaths.colorsAndFonts}`
-    }, {
-      name: TopLevelMenuItemName.icons,
-      link: `/${SegmentPaths.designer}/${SegmentPaths.iconsDeprecated}`
-    }];
   }
 
 }
